refactor(character-sheet): decompose into panels + WotC-inspired grid (#45)
Extract CharacterSheet.tsx (1177 lines) into composable panel components:
CharacterSheetHeader, AbilityScoresPanel, SavingThrowsPanel, CombatPanel,
FeaturesPanel, EquipmentPanel, SpellcastingPanel, PersonalityPanel,
BackstoryPanel + shared SectionHeader. Move EditHeader/Personality/Text and
the confirm dialog into character-sheet/edit-dialogs/.

Add a WotC-inspired named-area CSS grid (.sheet-grid: stats / combat / gear),
ability 'shield' boxes (large modifier + small score pill), and saving-throw
proficiency dots. Behaviour preserved: autosave, level controls, rest, clone/
archive/delete, conditions, hit-dice/spell-slot panels, portrait, empty/error
states all unchanged.
